test: move html tests in separate file
